Added non functional filtermenu
